refactor csharp classes to be used by both create from run and create blank unit test
