+++ Localize/lang/strings.json
@@ -1,6 +1,7 @@
 {
   "++ZVe/": "Testing",
   "+0H8Or": "Warning: input node type does not match the schema node's type",
+  "+0ua83": "Parameters",
   "+0yxlR": "Function display",
   "+3rROX": "Protected",
   "+5Jp42": "By",
@@ -663,6 +664,7 @@
   "Lub7NN": "Required. The expressions that may be true.",
   "LvLksz": "Loading outputs",
   "Lx8HRl": "(UTC+02:00) Damascus",
+  "LxRzQm": "Assertions",
   "M/gUE8": "About",
   "M/m3nG": "Action timeout",
   "M0bbZk": "Flow Checker",
@@ -802,6 +804,7 @@
   "QGbUXX": "Status code",
   "QHdKnm": "Drag and connect nodes to transform",
   "QNfUf/": "Full screen",
+  "QQmbz+": "Save unit test definition",
   "QT4IaP": "Filtered!",
   "QVtqAn": "Description",
   "QZBPUx": "Returns a single value matching the key name from form-data or form-encoded trigger output",
@@ -897,6 +900,7 @@
   "U3guPp": "Group by Connector",
   "U3iWVd": "Generates an array of integers starting from a certain number",
   "U4zovj": "Runs {onTime} {onDays}",
+  "U5O4Tz": "Create blank unit test (Preview)",
   "U7UAV0": "Mocked Results Tab",
   "U9SHxw": "Code",
   "UCNM4L": "To reference a parameter, use the dynamic content list.",
@@ -1080,9 +1084,11 @@
   "Zi9gQK": "Add new item",
   "ZihyUf": "Close",
   "ZkjTbp": "Learn more about dynamic content.",
+  "ZvAp7m": "Save",
   "ZyDq4/": "Show a different suggestion",
   "_++ZVe/.comment": "Title for testing section",
   "_+0H8Or.comment": "Warning message for when input node type does not match schema node type",
+  "_+0ua83.comment": "Button text for parameters",
   "_+0yxlR.comment": "Label for the function display radio group",
   "_+3rROX.comment": "Label in the chatbot header stating that the users information is protected in this chatbot",
   "_+5Jp42.comment": "Title for publisher",
@@ -1745,6 +1751,7 @@
   "_Lub7NN.comment": "Required expression parameters to apply or function",
   "_LvLksz.comment": "Loading outputs text",
   "_Lx8HRl.comment": "Time zone value ",
+  "_LxRzQm.comment": "Button text for unit test asssertions",
   "_M/gUE8.comment": "The tab label for the about tab on the operation panel",
   "_M/m3nG.comment": "title for action timeout setting",
   "_M0bbZk.comment": "Header for the errors panel",
@@ -1884,6 +1891,7 @@
   "_QGbUXX.comment": "Response status code for test map API",
   "_QHdKnm.comment": "default placeholder text",
   "_QNfUf/.comment": "Full Screen token picker",
+  "_QQmbz+.comment": "Button text for save unit test definition",
   "_QT4IaP.comment": "Filtered text",
   "_QVtqAn.comment": "Label for description column.",
   "_QZBPUx.comment": "Label for description of custom triggerFormDataValue Function",
@@ -1979,6 +1987,7 @@
   "_U3guPp.comment": "Label for the checkbox to group results by connector",
   "_U3iWVd.comment": "Label for description of custom range Function",
   "_U4zovj.comment": "Recurrence schedule description on days of week at times",
+  "_U5O4Tz.comment": "Button test for save blank unit test",
   "_U7UAV0.comment": "An accessibility label that describes the mocked results tab",
   "_U9SHxw.comment": "Code view title",
   "_UCNM4L.comment": "Description for Workflow Parameters Part 2",
@@ -2162,6 +2171,7 @@
   "_Zi9gQK.comment": "Label to add item to property editor",
   "_ZihyUf.comment": "Label for the close button in the chatbot header",
   "_ZkjTbp.comment": "Text for dynamic content link",
+  "_ZvAp7m.comment": "Button text for save",
   "_ZyDq4/.comment": "Text for the show different suggestion flow button",
   "_a7j3gS.comment": "Required number parameter to divide in mod function",
   "_aAXnqw.comment": "Required number of occurrences to get nthIndexOf function with",
@@ -2603,6 +2613,7 @@
   "_ocW+RF.comment": "Title for the details section in the template overview tab",
   "_odQ554.comment": "Response body for test map API",
   "_og5JOA.comment": "Millisecond",
+  "_ohOaXj.comment": "Button text for errors",
   "_ohpbkw.comment": "title for retry policy exponential interval setting",
   "_onXUu0.comment": "Text to tell users to click to add comments",
   "_oqgNX3.comment": "Accessibility label for workflow name",
@@ -2701,6 +2712,7 @@
   "_sFwHQc.comment": "aria label description for cancel button",
   "_sJQee6.comment": "Hours",
   "_sKy720.comment": "Error message when the workflow name is empty.",
+  "_sOnphB.comment": "Button text for resubmit",
   "_sRpETS.comment": "Warning message for when custom value does not match schema node type",
   "_sVQe34.comment": "The description for the test tab parameters.",
   "_sVcvcG.comment": "The tab label for the monitoring name and state tab on the create workflow panel",
@@ -2719,6 +2731,7 @@
   "_sv+IcU.comment": "Message to display when the data map definition can't be generated",
   "_svaqnp.comment": "Error message for when status is succeded and error is provided",
   "_sw6EXK.comment": "The title of the status property in the static result schema",
+  "_syPKiG.comment": "Button text for create unit test",
   "_syiNc+.comment": "Browse for file",
   "_sytRna.comment": "Error message to show when loading dynamic inputs failed",
   "_t+XCkg.comment": "Required string parameter for destination time zone",
@@ -3395,6 +3408,7 @@
   "ocW+RF": "Details",
   "odQ554": "Response body",
   "og5JOA": "{count} Millisecond",
+  "ohOaXj": "Errors",
   "ohpbkw": "Exponential interval",
   "onXUu0": "Add a note",
   "oqgNX3": "Workflow name",
@@ -3493,6 +3507,7 @@
   "sFwHQc": "Cancel creating a connection",
   "sJQee6": "{count} Hours",
   "sKy720": "Must provide value for workflow name.",
+  "sOnphB": "Resubmit",
   "sRpETS": "Warning: custom value does not match the schema node's type",
   "sVQe34": "Provide parameters to test the output.",
   "sVcvcG": "Basics",
@@ -3511,6 +3526,7 @@
   "sv+IcU": "Unable to generate data map definition",
   "svaqnp": "Error should not be provided when status is \"Succeeded\"",
   "sw6EXK": "Status",
+  "syPKiG": "Create unit test from run (Preview)",
   "syiNc+": "Browse",
   "sytRna": "Failed to retrieve dynamic inputs. Error details: ''{message}''",
   "t+XCkg": "Required. A string that contains the time zone name of the destination time zone. See https://msdn.microsoft.com/en-us/library/gg154758.aspx for details.",
